Refactoring and remove unused dependencies
