+++ internal_pages/embed_watermark_page.py
@@ -6,9 +6,9 @@ import time
 import os
 from modules.watermarking import embed_watermark
 from modules.face_recognition import get_face_embedding, check_face_match
-from modules.encryption import generate_key_with_helper, encrypt_watermark
-from modules.fuzzy_extractor import save_helper_data, serialize_helper_data
-from modules.utils import has_face, convert_to_bytes
+from modules.encryption import encrypt_watermark
+from modules.fuzzy_extractor import generate_key_with_helper
+from modules.utils import has_face, convert_to_bytes, save_helper_data
 from modules.constants import MAX_IMAGES, MAX_WATERMARK_SIZE
 
 def display_embed_watermark_page(debug_mode=False):
