+++ modules/fuzzy_extractor.py
@@ -1,9 +1,8 @@
 import numpy as np
-import pickle
 import os
 import hashlib
-import io
 from typing import Tuple, Optional, Dict, Any
+from modules.utils import bytes_to_bits, bits_to_bytes
 
 class FuzzyExtractor:
     """
@@ -32,7 +31,7 @@ class FuzzyExtractor:
             Tuple of (key, helper_data)
         """
         # Convert bytes to bit array
-        bits = self._bytes_to_bits(binary_data)
+        bits = bytes_to_bits(binary_data)
 
         # Create secure hash of original data as the key
         key = hashlib.sha256(binary_data).digest()
@@ -78,7 +77,7 @@ class FuzzyExtractor:
                 binary_data = binary_data + b'\x00' * (helper_data['data_length'] - len(binary_data))
 
         # Convert to bit array
-        bits = self._bytes_to_bits(binary_data)
+        bits = bytes_to_bits(binary_data)
         syndrome = np.array(helper_data['syndrome'], dtype=np.uint8)
 
         # Check data dimensions
@@ -109,7 +108,7 @@ class FuzzyExtractor:
         reconstructed_bits = syndrome ^ random_bits
 
         # Convert bits back to bytes
-        reconstructed_bytes = self._bits_to_bytes(reconstructed_bits)
+        reconstructed_bytes = bits_to_bytes(reconstructed_bits)
 
         # Verify reconstruction with hash check
         reconstruct_hash = hashlib.sha256(reconstructed_bytes).hexdigest()[:16]
@@ -143,7 +142,7 @@ class FuzzyExtractor:
             bits[i] = 1 - original_bit
 
             # Convert bits to bytes and check hash
-            test_bytes = self._bits_to_bytes(bits)
+            test_bytes = bits_to_bytes(bits)
             test_hash = hashlib.sha256(test_bytes).hexdigest()[:16]
 
             if test_hash == helper_data['hash_check']:
@@ -154,45 +153,7 @@ class FuzzyExtractor:
 
         return False
 
-    def _bytes_to_bits(self, data: bytes) -> np.ndarray:
-        """
-        Convert bytes to bit array.
-
-        Args:
-            data: Bytes to convert
-
-        Returns:
-            Numpy array of bits (0s and 1s)
-        """
-        result = []
-        for byte in data:
-            # Convert each byte to 8 bits
-            for i in range(8):
-                result.append((byte >> (7 - i)) & 1)
-        return np.array(result, dtype=np.uint8)
-
-    def _bits_to_bytes(self, bits: np.ndarray) -> bytes:
-        """
-        Convert bit array to bytes.
-
-        Args:
-            bits: Numpy array of bits (0s and 1s)
-
-        Returns:
-            Bytes representation
-        """
-        # Ensure length is multiple of 8
-        if len(bits) % 8 != 0:
-            bits = np.pad(bits, (0, 8 - (len(bits) % 8)))
-
-        result = bytearray()
-        for i in range(0, len(bits), 8):
-            byte_val = 0
-            for bit_idx, bit in enumerate(bits[i:i+8]):
-                byte_val |= (bit << (7 - bit_idx))
-            result.append(byte_val)
-
-        return bytes(result)
+    # Removed redundant wrapper methods - using imported functions directly
 
 
 def vector_to_binary(embedding: np.ndarray, threshold: float = 0.0) -> bytes:
@@ -211,24 +172,13 @@ def vector_to_binary(embedding: np.ndarray, threshold: float = 0.0) -> bytes:
         threshold = np.mean(embedding)
 
     # Create binary representation (1 for values >= threshold, 0 otherwise)
-    binary = (embedding >= threshold).astype(int)
+    binary = (embedding >= threshold).astype(np.uint8)
+    
+    # Convert to bytes
+    return bits_to_bytes(binary)
 
-    # Convert to bytes (8 bits per byte)
-    result = bytearray()
-    for i in range(0, len(binary), 8):
-        chunk = binary[i:i+8]
-        # Pad with zeros if necessary
-        if len(chunk) < 8:
-            chunk = np.pad(chunk, (0, 8 - len(chunk)))
-        byte_val = 0
-        for bit_idx, bit in enumerate(chunk):
-            byte_val |= (bit << (7 - bit_idx))
-        result.append(byte_val)
-
-    return bytes(result)
-
-def create_fuzzy_extractor(embedding: np.ndarray,
-                           error_tolerance: float = 0.15) -> Tuple[bytes, Dict[str, Any]]:
+def generate_key_with_helper(embedding: np.ndarray,
+                             error_tolerance: float = 0.15) -> Tuple[bytes, Dict[str, Any]]:
     """
     Create a fuzzy extractor from a face embedding vector.
     This implementation can handle larger binary data sizes.
@@ -262,8 +212,8 @@ def create_fuzzy_extractor(embedding: np.ndarray,
     return key, helper_dict
 
 
-def reproduce_key_from_embedding(embedding: np.ndarray,
-                                 helper_dict: Dict[str, Any]) -> Optional[bytes]:
+def regenerate_key_from_helper(embedding: np.ndarray,
+                               helper_dict: Dict[str, Any]) -> Optional[bytes]:
     """
     Reproduce the key from a face embedding and helper data.
 
@@ -288,77 +238,10 @@ def reproduce_key_from_embedding(embedding: np.ndarray,
 
         # Reproduce key from helper data
         key = extractor.reproduce(binary_data, helper_dict)
-
-        return key
+        if key is None:
+            return b''
+        else:
+            return key
     except Exception as e:
         print(f"Key reproduction error: {str(e)}")
-        return None
-
-def serialize_helper_data(helper_dict: Dict[str, Any]) -> bytes:
-    """
-    Serialize helper data to bytes for storage.
-
-    Args:
-        helper_dict: Helper data dictionary
-
-    Returns:
-        Serialized helper data as bytes
-    """
-    buffer = io.BytesIO()
-    pickle.dump(helper_dict, buffer)
-    return buffer.getvalue()
-
-def deserialize_helper_data(helper_bytes: bytes) -> Dict[str, Any]:
-    """
-    Deserialize helper data from bytes.
-
-    Args:
-        helper_bytes: Serialized helper data
-
-    Returns:
-        Helper data dictionary
-    """
-    buffer = io.BytesIO(helper_bytes)
-    return pickle.load(buffer)
-
-def save_helper_data(helper_dict: Dict[str, Any], path: str) -> bool:
-    """
-    Save helper data to a file.
-
-    Args:
-        helper_dict: Helper data dictionary
-        path: Path to save the helper data
-
-    Returns:
-        True if saving was successful, False otherwise
-    """
-    try:
-        # Create directory if it doesn't exist
-        os.makedirs(os.path.dirname(path), exist_ok=True)
-
-        # Serialize and save
-        helper_bytes = serialize_helper_data(helper_dict)
-        with open(path, 'wb') as f:
-            f.write(helper_bytes)
-        return True
-    except Exception as e:
-        print(f"Error saving helper data: {str(e)}")
-        return False
-
-def load_helper_data(path: str) -> Optional[Dict[str, Any]]:
-    """
-    Load helper data from a file.
-
-    Args:
-        path: Path to the helper data file
-
-    Returns:
-        Helper data dictionary or None if loading fails
-    """
-    try:
-        with open(path, 'rb') as f:
-            helper_bytes = f.read()
-        return deserialize_helper_data(helper_bytes)
-    except Exception as e:
-        print(f"Error loading helper data: {str(e)}")
-        return None
+        return b''
